requirements.txt
torch
transformers
datasets
sentence-transformers
faiss-cpu
openai
fastapi
uvicorn
pandas
numpy